refactor(raw article): Remove raw article translation implementation
