+++ src/main/java/org/zanata/mt/backend/ms/internal/dto/MSTranslateArrayResponse.java
@@ -3,7 +3,6 @@ package org.zanata.mt.backend.ms.internal.dto;
 /**
  * @author Alex Eng <a href="mailto:aeng@redhat.com">aeng@redhat.com</a>
  */
-
 import java.io.Serializable;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
