--- src/main/java/org/zanata/mt/service/ArticleConverter.java
@@ -1,21 +1,0 @@
-package org.zanata.mt.service;
-
-import org.zanata.mt.article.ArticleContents;
-
-/**
- * Interface for different Article converter.
- *
- * @author Alex Eng <a href="mailto:aeng@redhat.com">aeng@redhat.com</a>
- */
-public interface ArticleConverter {
-
-    /**
-     * Convert article into {@link ArticleContents}
-     *
-     * @param article
-     * @param srcLocale
-     * @param transLocale
-     * @param backendID
-     */
-    ArticleContents extractArticle(String html);
-}
