--- src/main/webapp/script/script.js
@@ -1,34 +1,0 @@
-/**
- * bookmarklet script to be use in browser
- *
- * This function will download script/article-script.js from MT service,
- * and execute article-script.js#requestTranslations() after page is loaded.
- *
- * ${service.url} is set in pom.xml profile
- */
-function bookmarklet() {
-  var BASE_URL = '${service.url}';
-  var scriptUrl = BASE_URL + '/script/article-script.js';
-  var script = document.createElement('script');
-  script.setAttribute('src', scriptUrl);
-  script.addEventListener('load', function() {
-    requestTranslations();
-  }, false);
-  document.body.appendChild(script);
-}
-
-// To display raw js in index.jsp
-function displayRawJS() {
-  document.getElementById('raw-js').innerHTML = bookmarklet.toString();
-}
-
-/**
- * To display bookmarklet js in index.jsp
- *
- */
-function displayBookmarklet() {
-  var BASE_URL = '${service.url}';
-  var bookmarklet = "javascript:(function()%7Bvar%20BASE_URL%20%3D%20'" + encodeURIComponent(BASE_URL) + "'%3Bvar%20scriptUrl%20%3D%20BASE_URL%20%2B%20'%2Fscript%2Farticle-script.js'%3Bvar%20script%20%3D%20document.createElement('script')%3Bscript.setAttribute('src'%2C%20scriptUrl)%3Bscript.addEventListener('load'%2C%20function()%20%7BrequestTranslations()%3B%7D%2C%20false)%3Bdocument.body.appendChild(script)%7D)()"
-  document.getElementById('bookmarklet-link').href = bookmarklet;
-  document.getElementById('bookmarklet-area').value = bookmarklet;
-}
