+++ src/test/java/org/zanata/mt/api/InputStreamStreamingOutputTest.java
@@ -11,7 +11,6 @@ import java.io.OutputStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
