+++ src/test/java/org/zanata/mt/exception/ZanataMTExceptionTest.java
@@ -1,7 +1,6 @@
 package org.zanata.mt.exception;
 
 import org.junit.Test;
-import org.mockito.Mockito;
 
 import static org.assertj.core.api.Assertions.assertThat;
