+++ src/test/java/org/zanata/mt/util/DomUtilTest.java
@@ -1,8 +1,5 @@
 package org.zanata.mt.util;
 
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.select.Elements;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -13,39 +10,44 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class DomUtilTest {
 
     @Test
-    public void testParseAsElement() {
-        Elements elements = DomUtil.parseAsElement(
-                "<html><body><tag>content to return</tag></body></html>");
-        assertThat(elements).isNotNull().isNotEmpty();
+    public void privateNotesAreRecognised() {
+        String html = "<div id='private-notes-testing'></div>";
+        assertThat(DomUtil.isKCSPrivateNotes(html)).isTrue();
     }
 
     @Test
-    public void testParseAsElementEmptyBody() {
-        Elements elements = DomUtil.parseAsElement(
-                "<html></html>");
-        assertThat(elements).isNull();
+    public void nonPrivateNotesAreRecognised() {
+        String html = "<div id='testing-private-notes'></div>";
+        assertThat(DomUtil.isKCSPrivateNotes(html)).isFalse();
     }
 
     @Test
-    public void testParseAsElementEmptyBodyChildren() {
-        Elements elements = DomUtil.parseAsElement(
-                "<html><body></body></html>");
-        assertThat(elements).isNull();
+    public void codeSectionRecognised() {
+        String html = "<div id='code-raw'></div>";
+        assertThat(DomUtil.isKCSCodeSection(html)).isTrue();
     }
 
     @Test
-    public void testExtractEmptyBodyContentHTML() {
-        String html = "<html><body></body></html>";
-        Document doc = Jsoup.parse(html);
-        String bodyHtml = DomUtil.extractBodyContentHTML(doc);
-        assertThat(bodyHtml).isEqualTo("");
+    public void nonCodeSectionRecognised() {
+        String html = "<div id='not-code-raw'></div>";
+        assertThat(DomUtil.isKCSCodeSection(html)).isFalse();
     }
 
     @Test
-    public void testExtractBodyHTML() {
-        String html = "<html><body>test body</body></html>";
-        Document doc = Jsoup.parse(html);
-        String bodyHtml = DomUtil.extractBodyContentHTML(doc);
-        assertThat(bodyHtml).isEqualTo("test body");
+    public void isNonTranslatableNodeRecognised() {
+        String html = "<div id='code-raw' translate='no'></div>";
+        assertThat(DomUtil.isNonTranslatableNode(html)).isTrue();
+    }
+
+    @Test
+    public void translatableNodeRecognised() {
+        String html = "<div id='code-raw' translate='yes'></div>";
+        assertThat(DomUtil.isNonTranslatableNode(html)).isFalse();
+    }
+
+    @Test
+    public void translatableNodeRecognised2() {
+        String html = "<div id='code-raw'></div>";
+        assertThat(DomUtil.isNonTranslatableNode(html)).isFalse();
     }
 }
