--- src/test/webapp/demo/kcs_sample.html
@@ -1,38 +1,0 @@
-<html>
-  <head>
-    <meta name="language" content="en"/>
-    <title>Zanata KCS Article Test</title>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-  </head>
-  <body class="i18n-fr">
-  <div id="cp-content" class="main-content">
-    <article class="kcs_solution cirrus-content" id="main-content">
-      <div class="content-wrapper">
-        <header class="header">
-          <h1 class="title">English to French KCS test</h1>
-          <div class="header-meta">
-            <span class="status inprogress" data-content="This solution is in progress.">Solution In Progress</span>
-          </div>
-        </header>
-        <section>
-          <h2>Normal section</h2>
-          <p>This is normal section, should be translated</p>
-        </section>
-        <section>
-          <h2>Coding section</h2>
-          <div class="code-raw">
-            <pre>source code1, stack trace here. Should not be translated</pre>
-          </div>
-          <div class="code-raw">
-            <pre>source code2, stack trace here. Should not be translated</pre>
-          </div>
-        </section>
-        <section id="private-notes-section">
-          <h2>private section</h2>
-          <p>private notes which should be ignore</p>
-        </section>
-      </div>
-    </article>
-  </div>
-  </body>
-</html>
